+++ src/lib/components/general/ModalLib.svelte
@@ -1,17 +1,17 @@
 <script lang="ts">
 import { modalStore } from "$lib/stores/modal";
 import type { SupabaseClient } from "@supabase/supabase-js";
-import AdvancedSearch from "../recal/elements/adv/AdvancedSearch.svelte";
-import EditCb from "../recal/elements/cb/editCb.svelte";
-import ManageCb from "../recal/elements/cb/manageCb.svelte";
-import ExportCal from "../recal/elements/export/ExportCal.svelte";
-import ShareCal from "../recal/elements/export/ShareCal.svelte";
+import AdvancedSearch from "../recal/modals/AdvancedSearch.svelte";
+import EditCb from "../recal/modals/EditCb.svelte";
+import ManageCb from "../recal/modals/ManageCb.svelte";
+import ExportCal from "../recal/modals/ExportCal.svelte";
+import ShareCal from "../recal/modals/ShareCal.svelte";
 import AddSchedule from "../recal/elements/save/AddSchedule.svelte";
 import EditSchedule from "../recal/elements/save/EditSchedule.svelte";
 import Pinned from "../recal/elements/save/Pinned.svelte";
-import RecalColors from "../recal/elements/adv/RecalColors.svelte";
-import RecalPalettes from "../recal/elements/adv/RecalPalettes.svelte";
-    import RecalUtils from "../recal/elements/adv/RecalUtils.svelte";
+import RecalColors from "../recal/modals/RecalColors.svelte";
+import RecalPalettes from "../recal/modals/RecalPalettes.svelte";
+import RecalUtils from "../recal/modals/RecalUtils.svelte";
 
 export let supabase: SupabaseClient;
 
@@ -60,4 +60,3 @@ export let supabase: SupabaseClient;
 {#if $modalStore === "rutils"}
 <RecalUtils showModal={$modalStore === "rutils"} />
 {/if}
-
