--- src/lib/components/recal/elements/cards/MinimalBase.stories.ts
@@ -1,15 +1,0 @@
-import MinimalBase__SvelteComponent_ from "./MinimalBase.svelte";
-import { action } from "@storybook/addon-actions";
-
-export default {
-    component: MinimalBase__SvelteComponent_,
-    title: "Minimal Base",
-}
-
-const Template = ({ ...args }) => ({
-    Component: MinimalBase__SvelteComponent_,
-    props: args,
-});
-
-
-export const Default = Template.bind({});
